pyproject.toml
[tool.poetry]
name = "lonelyconnect"
version = "0.1.0"
description = "Make your own OnlyConnect-style quizzes at home"
authors = ["L3viathan <git@l3vi.de>"]
readme = "README.md"
repository = "https://github.com/L3viathan/lonelyconnect"
classifiers = [
    "Environment :: Web Environment",
    "Operating System :: OS Independent",
    "Topic :: Games/Entertainment :: Puzzle Games",
]

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^5.4.1"
fastapi = "^0.68.0"
Jinja2 = "^3.0.1"
MarkupSafe = "^2.0.1"
aiofiles = "^0.7.0"
python-multipart = "^0.0.5"
uvicorn = "^0.15.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
freezegun = "^1.1.0"

[tool.poetry.scripts]
lonelyconnect = "lonelyconnect:entrypoint"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
